[green] fix SSDTest link problem
