+++ SSDProject/SSDProject/SSDProject.vcxproj.filters
@@ -24,13 +24,7 @@
     <ClCompile Include="SSD.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="Command.cpp">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
     <ClCompile Include="StorageException.cpp">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
-    <ClCompile Include="Parser.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="StorageException.cpp">
